requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.0.0
requests==2.31.0